+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next Release
 
+### Changes
+- Add CVD risk view to patient summary screen behind feature flag `show_cvd_risk`
+
 ## 2026.06.25
 
 ### Changes